manejo de excepciones

+++ src/main/java/com/barberia/barberia/controllers/CitaController.java
@@ -1,6 +1,9 @@
 package com.barberia.barberia.controllers;
 
+import com.barberia.barberia.entities.Barbero;
 import com.barberia.barberia.entities.Cita;
+import com.barberia.barberia.exceptions.BarberoNoExisteException;
+import com.barberia.barberia.exceptions.FechaHoraPasadaException;
 import com.barberia.barberia.exceptions.HorarioNoDisponibleException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -8,12 +11,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import com.barberia.barberia.services.CitaService;
 
+import java.time.OffsetDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/citas")
 public class CitaController {
-
     private final CitaService citaService;
 
     @Autowired
@@ -28,24 +32,38 @@ public class CitaController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Cita> getCitaById(@PathVariable Integer id) {
-        return citaService.getCitaById(id)
-                .map(cita -> new ResponseEntity<>(cita, HttpStatus.OK))
-                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        Optional<Cita> cita = citaService.getCitaById(id);
+        return cita.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @PostMapping("/nueva")
-    public ResponseEntity<?> crearNuevaCita(@RequestBody Cita cita) {
+    @PostMapping("/crear")
+    public ResponseEntity<?> createCita(@RequestBody Cita cita) {
         try {
             Cita nuevaCita = citaService.saveCita(cita);
-            return new ResponseEntity<>(nuevaCita, HttpStatus.CREATED);
-        } catch (HorarioNoDisponibleException ex) {
-            return new ResponseEntity<>("El horario seleccionado no está disponible.", HttpStatus.BAD_REQUEST);
+            return ResponseEntity.ok(nuevaCita);
+        } catch (HorarioNoDisponibleException | BarberoNoExisteException | FechaHoraPasadaException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+    }
+    @PutMapping("/id")
+    public ResponseEntity<Cita> updateCita(@PathVariable Integer id, @RequestBody Cita updatedCita) {
+        Optional<Cita> existingCita = citaService.getCitaById(id);
+        if (existingCita.isPresent()) {
+            updatedCita.setId(id);
+            return ResponseEntity.ok(citaService.saveCita(updatedCita));
+        } else {
+            return ResponseEntity.notFound().build();
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCita(@PathVariable Integer id) {
-        citaService.deleteCita(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        Optional<Cita> existingCita = citaService.getCitaById(id);
+        if (existingCita.isPresent()) {
+            citaService.deleteCita(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
